feat: Enhance trash models with deletion tracking and user references
- Added `deletedAt` and `deletedBy` fields to various trash models (TrashCoupon, TrashEvent, TrashGuide, TrashNursery, TrashPartner, TrashReport, TrashResponse, TrashTree, TrashUser) to track deletion timestamps and the user responsible for the deletion.
- Updated the TrashTask model by removing it as it is no longer needed.
- Modified services (CouponService, EventService, GuideService, NurseryService, PartnerService, ReportService, ResponseService, UserService) to handle deletion logic, including moving items to trash and managing related data.
- Updated validators for events, guides, nurseries, reports, and trees to ensure proper data structure and validation.
- Refactored comment and report deletion logic to include handling of replies and responses.

+++ Backend/src/controllers/responseController.ts
@@ -72,7 +72,8 @@ export default class ResponseController {
      * @access    Private
     */
     deleteResponse = asyncHandler(async (req: Request, res: Response, next: NextFunction) => {
-        const result = await this.ResponseService.deleteResponse(req.params.id);
+        const { id, role } : {id: string, role: string} = req.body.user;
+        const result = await this.ResponseService.deleteResponse(req.params.id, {role, id});
         if (result.status === 200) {
             res.json({ message: result.message });
         } else {
